Changed parser elements matched
The parser was aggressively matching single elements that most likely
were not spell combinations.  For example, the following sentence would
be matched:  A weapon starts with W.  The A and W would both be
converted.  I changed this so only the letter "E" is matched if it is
single, all other elements including X won't convert if they are alone.
I've also updated both the extensions and tested them but I have no idea
if they will upload to the stores correctly as I've never made a chrome
or firefox extension before.

Check index.html for more details.